requirements.txt
pandas==1.3.1
beautifulsoup4==4.9.3
Flask==1.1.1
slackclient==2.9.3
requests==2.22.0
